requirements.txt
aiomysql==0.3.0
aioredis==2.0.1
aiosmtplib==5.1.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncmy==0.2.11
asyncpg==0.31.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.129.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastapi-mail==1.6.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg2==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
PyMySQL==1.1.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.1.1
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.2.0
regex==2025.11.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# Advanced Features
celery==5.4.0
slowapi==0.1.9
strawberry-graphql==0.254.0
python-multipart==0.0.22
aiofiles==24.1.0
authlib==1.3.2
itsdangerous==2.2.0

